% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{all_equal_S}
\alias{all_equal_S}
\title{all.equal with Sign Alignment for Matrices}
\usage{
all_equal_S(A, B)
}
\arguments{
\item{A}{A numerical matrix of dimension \code{n}-times-\code{d}.}

\item{B}{A numerical matrix of dimension \code{n}-times-\code{d}.}
}
\value{
A result of function \link[base:all.equal]{all.equal} applied to
aligned matrices \code{A} and \code{B}. Gives \code{TRUE} if the two matrices
are almost equal, up to the signs of their columns.
}
\description{
Function \link[base:all.equal]{all.equal} that is adjusted so that possible
differences in sign between the columns of the two matrices are not taken 
into account.
}
\examples{
A = matrix(c(1,0,0, 0,1,0, 0,0,1),ncol=3, byrow=TRUE)
B = matrix(c(-1,0,0, 0,-1,0, 0,0,1),ncol=3, byrow=TRUE)

all.equal(A, B)
all_equal_S(A, B)
}
\seealso{
\link[base:all.equal]{all.equal}
}
