% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the location of accidents in a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer that represents the state number.}

\item{year}{An integer or string indicating the year.}
}
\value{
Plots a map of the state, where the fatalities are the dots on the map.
}
\description{
Plot the location of accidents in a given state and year
}
\details{
This function takes the state number and a year as inputs and calls the supporting
   functions \code{make_filename()} and \code{fars_read()}.
}
\examples{
\dontrun{
setwd('data')
fars_map_state(12, 2013)
}
}
