% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing the name of the csv file to be read,
which must be at the CURRENT directory.}
}
\value{
If the file exists, this function returns a data frame table. If the file doesn't
   exist it raises an error message.
}
\description{
Read data from csv file
}
\examples{
\dontrun{
fars_read(file.path("data", "accident_2013.csv.bz2")) # file in current working directory
fars_read("data123.csv") # Error: file 'data123.csv' does not exist.
}

}
