% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A list of FARS data for the years in the parameter \code{years}.
}
\description{
Reads data for multiple years
}
\examples{
\dontrun{
fars_read_years(c(2014, 2015, 2016))
}
}
