% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data data for multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A dataframe with monthly number of accidents in each row and years as colunms.
}
\description{
Summarize data data for multiple years
}
\details{
This function produces a "tidy" data frame of FARS data for multiple years. Data
   is summarized by each month of each year with each year in its own column. This function
   uses \code{\link{fars_read_years}} function to generate a list of month and year.
}
\examples{
fars_summarize_years(c(2014, 2015))

}
