% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{MAD_R}
\alias{MAD_R}
\title{Median Absolute Deviation}
\usage{
MAD_R(x)
}
\arguments{
\item{x}{A numerical vector whose MAD we compute.}
}
\value{
A single numerical value, median of the absolute distances from the 
median of \code{x}.
}
\description{
Standard median absolute deviation (MAD) of a numerical vector. Compared to
function \link[stats:mad]{mad}, no additional adjustment factor for 
consistency with normal distribution is used.
}
\examples{
n = 10
X = rnorm(n)

MAD_R(X)
mad(X)/1.4826 # mad is using an adjustment factor 1.4826
}
\seealso{
\link[stats:mad]{mad}
}
