% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate the name of the file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or string indicating the year that generate the data file.}
}
\description{
Generate the name of the file
}
\details{
This function uses the year (input) to generate the file name for the Fatality
   Analysis Reporting System (FARS) data file. It converts the year into an integer and
   then uses the \code{sprintf()} function to format the file name.
}
\examples{
\dontrun{
make_filename(2014) #generates "accident_2014.csv.bz2"
}
}
