\name{RPD-package}
\alias{RPD-package}
\alias{RPD}
\docType{package}
\title{
\packageTitle{RPD}
}
\description{
\packageDescription{RPD}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{RPD}

Maintainer: \packageMaintainer{RPD}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = RPD)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{RPD}}
%\section{Documentation Index}{\packageIndices{RPD}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
